🎨 Palette: improve navbar accessibility and fix CI/lint errors
- Add 'Skip to content' link for keyboard users
- Upgrade navigation to use NavLink with active state indicators
- Make Navbar sticky for constant accessibility
- Fix apostrophe ESLint errors in Navbar and ErrorBoundary
- Add id='main-content' to App.tsx
- Remove junk workspace file
- Update App.test.tsx to verify new features
- Sync package-lock.json for CI
- Add continue-on-error to Snyk security scan in deploy.yml
- Fix CodeQL shell injection vulnerability in gitflow action
- Resolve 15+ ESLint errors (any types, undefined globals, Fast Refresh warnings)
- Move useToast hook to separate file src/hooks/useToast.ts
