+++ scripts/test-openrouter.mjs
@@ -1,3 +1,4 @@
+/* global process */
 import fetch from 'node-fetch';
 import dotenv from 'dotenv';
 import path from 'path';
