+++ src/pages/CreatePost.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { generateBlogPost, analyzeContent } from '../lib/openrouter'
 import { createPost } from '../lib/supabase'
-import { useToast } from '../components/ui/Toaster'
+import { useToast } from '../hooks/useToast'
 import { formatToMarkdown } from '../utils/markdown'
 import ReactMarkdown from 'react-markdown'
 import remarkGfm from 'remark-gfm'
