+++ tailwind.config.js
@@ -1,3 +1,4 @@
+/* global require */
 /** @type {import('tailwindcss').Config} */
 export default {
   content: [
